break: phase-out serializeToModule
move serializeToModule into separate package serialize-to-module

docu: fix md
