+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "serialize-to-js",
-  "version": "2.0.1",
+  "version": "3.0.0-0",
   "description": "serialize objects to javascript",
   "keywords": [
     "javascript",
@@ -13,21 +13,20 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/commenthol/serialize-to-js.git"
+    "url": "git+https://github.com/commenthol/serialize-to-js.git"
   },
   "license": "MIT",
   "author": "commenthol <commenthol@gmail.com>",
   "main": "lib",
   "directories": {
-    "doc": "doc",
+    "lib": "lib",
     "test": "test"
   },
   "scripts": {
     "all": "npm run lint && npm test",
     "clean": "rimraf doc coverage .nyc_output node_modules *.tgz",
     "coverage": "nyc -r text -r html npm test",
-    "doc": "jsdox -o doc lib/*.js",
-    "lint": "eslint '**/*.js'",
+    "lint": "eslint lib test",
     "prepublishOnly": "npm run all",
     "readme": "markedpp --githubid -i README.md -o README.md",
     "test": "mocha"
@@ -41,14 +40,12 @@
       "key-spacing": 0
     }
   },
-  "dependencies": {
-    "js-beautify": "^1.10.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "eslint": "^5.16.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-node": "^9.0.1",
+    "eslint-plugin-node": "^9.1.0",
     "eslint-plugin-promise": "^4.1.1",
     "eslint-plugin-standard": "^4.0.0",
     "mocha": "^6.1.4",
